[CLEANUP] Move TestCase helpers to internal-test-helpers
