--- packages/ember-glimmer/tests/utils/environment.js
@@ -1,1 +1,0 @@
-export { Environment as default } from 'ember-glimmer';
