+++ packages/ember-glimmer/tests/utils/test-case.js
@@ -1,21 +1,7 @@
-export { TestCase, moduleFor } from './abstract-test-case';
-import {
-  AbstractApplicationTest,
-  AbstractRenderingTest
-} from './abstract-test-case';
-import { ComponentLookup } from 'ember-views';
+export {
+  AbstractTestCase as TestCase,
+  ApplicationTestCase as ApplicationTest,
+  RenderingTestCase as RenderingTest,
+  moduleFor
+} from 'internal-test-helpers';
 
-export class ApplicationTest extends AbstractApplicationTest {
-}
-
-export class RenderingTest extends AbstractRenderingTest {
-  constructor() {
-    super();
-    let { owner } = this;
-
-    this.env = owner.lookup('service:-glimmer-environment');
-    owner.register('component-lookup:main', ComponentLookup);
-    owner.registerOptionsForType('helper', { instantiate: false });
-    owner.registerOptionsForType('component', { singleton: false });
-  }
-}
