+++ packages/internal-test-helpers/lib/index.js
@@ -1,6 +1,17 @@
 export { default as factory } from './factory';
 export { default as buildOwner } from './build-owner';
 export { default as confirmExport } from './confirm-export';
+export { default as equalInnerHTML } from './equal-inner-html';
+export { default as equalTokens } from './equal-tokens';
+export { default as moduleFor } from './module-for';
+export { default as strip } from './strip';
+export { default as applyMixins } from './apply-mixins';
+export {
+  equalsElement,
+  classes,
+  styles,
+  regex
+} from './matchers';
 export {
   runAppend,
   runDestroy
@@ -9,3 +20,9 @@ export {
   testBoth,
   testWithDefault
 } from './test-groups';
+
+export { default as AbstractTestCase } from './test-cases/abstract';
+export { default as AbstractApplicationTestCase } from './test-cases/abstract-application';
+export { default as ApplicationTestCase } from './test-cases/application';
+export { default as AbstractRenderingTestCase } from './test-cases/abstract-rendering';
+export { default as RenderingTestCase } from './test-cases/rendering';
