+++ packages/internal-test-helpers/lib/module-for.js
@@ -1,0 +1,32 @@
+import applyMixins from './apply-mixins';
+
+export default function moduleFor(description, TestClass, ...mixins) {
+  let context;
+
+  QUnit.module(description, {
+    setup() {
+      context = new TestClass();
+    },
+
+    teardown() {
+      context.teardown();
+    }
+  });
+
+  applyMixins(TestClass, mixins);
+
+  let proto = TestClass.prototype;
+
+  while (proto !== Object.prototype) {
+    Object.keys(proto).forEach(generateTest);
+    proto = Object.getPrototypeOf(proto);
+  }
+
+  function generateTest(name) {
+    if (name.indexOf('@test ') === 0) {
+      QUnit.test(name.slice(5), assert => context[name](assert));
+    } else if (name.indexOf('@skip ') === 0) {
+      QUnit.skip(name.slice(5), assert => context[name](assert));
+    }
+  }
+}
