+++ packages/internal-test-helpers/lib/strip.js
@@ -1,0 +1,7 @@
+export default function strip([...strings], ...values) {
+  let str = strings.map((string, index) => {
+    let interpolated = values[index];
+    return string + (interpolated !== undefined ? interpolated : '');
+  }).join('');
+  return str.split('\n').map(s => s.trim()).join('');
+}
