+++ packages/internal-test-helpers/lib/test-cases/application.js
@@ -1,0 +1,4 @@
+import AbstractApplicationTestCase from './abstract-application';
+
+export default class ApplicationTestCase extends AbstractApplicationTestCase {
+}
